refactor and add separate hashing functions
